refactor: 코드 리뷰 내용 적용
